+++ ecommerce_final/order-service/src/main/kotlin/it/polito/wa2/orderservice/schedulingTasks/ScheduledTask.kt
@@ -33,7 +33,7 @@ class ScheduledTask(
     /**
      * Scheduler to remove the finished sagas and order job from the list and free memory
      */
-    @Scheduled(fixedRate = 1000*10) // 10s
+    @Scheduled(fixedRate = (1000*10).toLong()) // 10s
     fun removeCompletedSagas() = CoroutineScope(Dispatchers.Default).launch{
         val sagas = getListOfStateMachine()
         logger.info("BEFORE REMOVED SAGAS: $sagas --- REMOVED JOBS: $jobs")
